refactor gas pricing
