+++ src/ethereum/prague/vm/gas.py
@@ -20,7 +20,7 @@ from ethereum.trace import GasAndRefund, evm_trace
 from ethereum.utils.numeric import ceil32, taylor_exponential
 
 from ..blocks import Header
-from ..transactions import BlobTransaction, Transaction
+from ..transactions import Transaction
 from . import Evm
 from .exceptions import OutOfGasError
 
@@ -67,11 +67,6 @@ GAS_WARM_ACCESS = Uint(100)
 GAS_INIT_CODE_WORD_COST = Uint(2)
 GAS_BLOBHASH_OPCODE = Uint(3)
 GAS_POINT_EVALUATION = Uint(50000)
-
-TARGET_BLOB_GAS_PER_BLOCK = U64(786432)
-GAS_PER_BLOB = U64(2**17)
-MIN_BLOB_GASPRICE = Uint(1)
-BLOB_BASE_FEE_UPDATE_FRACTION = Uint(5007716)
 
 GAS_BLS_G1_ADD = Uint(375)
 GAS_BLS_G1_MUL = Uint(12000)
@@ -258,113 +253,18 @@ def max_message_call_gas(gas: Uint) -> Uint:
     return gas - (gas // Uint(64))
 
 
-def init_code_cost(init_code_length: Uint) -> Uint:
+def init_code_cost(length: Uint) -> Uint:
     """
-    Calculates the gas to be charged for the init code in CREAT*
-    opcodes as well as create transactions.
+    Calculate the gas cost for init code.
 
     Parameters
     ----------
-    init_code_length :
-        The length of the init code provided to the opcode
-        or a create transaction
+    length :
+        The length of the init code.
 
     Returns
     -------
-    init_code_gas: `ethereum.base_types.Uint`
-        The gas to be charged for the init code.
+    gas_cost : `ethereum.base_types.Uint`
+        The gas cost for the init code.
     """
-    return GAS_INIT_CODE_WORD_COST * ceil32(init_code_length) // Uint(32)
-
-
-def calculate_excess_blob_gas(parent_header: Header) -> U64:
-    """
-    Calculated the excess blob gas for the current block based
-    on the gas used in the parent block.
-
-    Parameters
-    ----------
-    parent_header :
-        The parent block of the current block.
-
-    Returns
-    -------
-    excess_blob_gas: `ethereum.base_types.U64`
-        The excess blob gas for the current block.
-    """
-    # At the fork block, these are defined as zero.
-    excess_blob_gas = U64(0)
-    blob_gas_used = U64(0)
-
-    if isinstance(parent_header, Header):
-        # After the fork block, read them from the parent header.
-        excess_blob_gas = parent_header.excess_blob_gas
-        blob_gas_used = parent_header.blob_gas_used
-
-    parent_blob_gas = excess_blob_gas + blob_gas_used
-    if parent_blob_gas < TARGET_BLOB_GAS_PER_BLOCK:
-        return U64(0)
-    else:
-        return parent_blob_gas - TARGET_BLOB_GAS_PER_BLOCK
-
-
-def calculate_total_blob_gas(tx: Transaction) -> U64:
-    """
-    Calculate the total blob gas for a transaction.
-
-    Parameters
-    ----------
-    tx :
-        The transaction for which the blob gas is to be calculated.
-
-    Returns
-    -------
-    total_blob_gas: `ethereum.base_types.Uint`
-        The total blob gas for the transaction.
-    """
-    if isinstance(tx, BlobTransaction):
-        return GAS_PER_BLOB * U64(len(tx.blob_versioned_hashes))
-    else:
-        return U64(0)
-
-
-def calculate_blob_gas_price(excess_blob_gas: U64) -> Uint:
-    """
-    Calculate the blob gasprice for a block.
-
-    Parameters
-    ----------
-    excess_blob_gas :
-        The excess blob gas for the block.
-
-    Returns
-    -------
-    blob_gasprice: `Uint`
-        The blob gasprice.
-    """
-    return taylor_exponential(
-        MIN_BLOB_GASPRICE,
-        Uint(excess_blob_gas),
-        BLOB_BASE_FEE_UPDATE_FRACTION,
-    )
-
-
-def calculate_data_fee(excess_blob_gas: U64, tx: Transaction) -> Uint:
-    """
-    Calculate the blob data fee for a transaction.
-
-    Parameters
-    ----------
-    excess_blob_gas :
-        The excess_blob_gas for the execution.
-    tx :
-        The transaction for which the blob data fee is to be calculated.
-
-    Returns
-    -------
-    data_fee: `Uint`
-        The blob data fee.
-    """
-    return Uint(calculate_total_blob_gas(tx)) * calculate_blob_gas_price(
-        excess_blob_gas
-    )
+    return length * GAS_INIT_CODE_WORD_COST
